chore: move typeguards to separate file
